Updated wizard on join organization component
The wizard for the `joinorg` component has been designed inside of the component.
Since rules for each step will be different for the join and create organization components,
they both will be created inside of the component instead of from a generic wizard.
Images of this wizard will be included in the PR comments...